styling and file organization
